Refactor classes into their own files
Former-commit-id: c57acf3335ca989764bff48fb35ff03070b3fc66
