+++ lib/rack/cors/resources/cors_misconfiguration_error.rb
@@ -1,0 +1,12 @@
+module Rack
+  class Cors
+    class Resource
+      class CorsMisconfigurationError < StandardError
+        def message
+          "Allowing credentials for wildcard origins is insecure." \
+          " Please specify more restrictive origins or set 'credentials' to false in your CORS configuration."
+        end
+      end
+    end
+  end
+end
